refactor(samples): split uploadDirectory sample to a new file (#829)
